Added transition argument: 'random'
A Transition option set to 'random' will randomly choose a transition
for each slide before it is shown. (The demo currently throws an error
if a CSS transition is selected, though I believe that the
modifications are sound and will work as intended when using either CSS
or Javascript Transitions.